backend/requirements.txt
alabaster==0.7.16
asgiref==3.8.1
attrs==23.2.0
Babel==2.15.0
cattrs==23.2.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.6
django-cors-headers==4.3.1
django-netfields==1.3.2
django-schema-graph==3.1.0
djangorestframework==3.15.1
docutils==0.21.2
drf-spectacular==0.27.2
exceptiongroup==1.2.1
idna==3.7
imagesize==1.4.1
inflection==0.5.1
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
netaddr==1.2.1
packaging==24.0
psycopg==3.1.19
psycopg-binary==3.1.19
Pygments==2.18.0
PyYAML==6.0.1
referencing==0.35.1
requests==2.31.0
rpds-py==0.18.1
six==1.16.0
snowballstemmer==2.2.0
Sphinx==7.3.7
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
